Data module rework
Reworked data module to use config files instead of saving data directly as a resource, because concerns about easy-to-perform ACE got the best of me.
As such, added two methods to the clue class in order to help with saving them to file (a convert-to-dictionary method and a convert-from-dictionary method).